Changing int to double in prices. Fixing bugs and implementing get service offer on front-end.

+++ backend/SEP/AgencyService/DTO/ServiceOfferItemDto.cs
@@ -7,7 +7,7 @@ namespace AgencyService.DTO
         public int Id { get; set; }
         public EOfferName OfferName { get; set; }
         public bool IsAccepted { get; set; }
-        public int MonthlyPrice { get; set; }
-        public int YearlyPrice { get; set; }
+        public double MonthlyPrice { get; set; }
+        public double YearlyPrice { get; set; }
     }
 }
